[chat] - fix: remove race condition by removing buffer from pipeStream channels
The handling of events meant that there could be a race condition, where a select could
see the EOF before the last token (when multiple go channels are ready in a select, it
chooses one randomly). In fact, this behavior was pretty consistent with the latest version
of sgpt, we were nearly always missing the last batch of tokens from openai.

Removing the buffers means that now, tokens must be handled before the error, forcing the
order of events and guaranteeing to see the last tokens before the EOF.